+++ src/pages/create-post/ui/CreatePostPage.tsx
@@ -5,11 +5,13 @@ import { useState } from 'react'
 
 import { ROUTES } from '@/shared/config'
 
+import { useCreatePostFlow } from '../model/useCreatePostFlow'
 import { CreatePostModal } from './CreatePostModal'
 
 export const CreatePostPage = () => {
   const router = useRouter()
   const [isOpen, setIsOpen] = useState(true)
+  const { photos, selectPhotosHandler, uploadError } = useCreatePostFlow()
 
   const closeHandler = (open: boolean) => {
     setIsOpen(open)
@@ -19,5 +21,13 @@ export const CreatePostPage = () => {
     }
   }
 
-  return <CreatePostModal open={isOpen} onOpenChange={closeHandler} />
+  return (
+    <CreatePostModal
+      open={isOpen}
+      onOpenChange={closeHandler}
+      onPhotosSelect={selectPhotosHandler}
+      photos={photos}
+      uploadError={uploadError}
+    />
+  )
 }